requirements.txt
attrs
pypika